dgfit models introduced - allows for different size distribution models (MRN, WG, bins)
